# docs/swagger.yaml
basePath: /api
definitions:
...
  user.Profile:
    properties:
      avatar_url:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  user.certifyReq:
    properties:
      company_name:
        type: string
      position:
        type: string
    required:
    - company_name
    - position
    type: object
  user.updateProfileReq:
    properties:
      avatar_url:
        type: string
      name:
        type: string
    required:
    - name
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/maxfeizi04-cloud/recruitment-platform
  description: 放心招聘平台后端接口文档，支持求职者、招聘者、管理员三端
  title: 放心招聘平台 API
  version: "2.0"
paths:
...
  /hrs/certification:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取HR认证信息
      tags:
      - 用户
  /hrs/certify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 公司名称和职位
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.certifyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 提交HR认证
      tags:
      - 用户
...
  /users/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Profile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取个人资料
      tags:
      - 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: 姓名和头像
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.updateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Profile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新个人资料
      tags:
      - 用户
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"